# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: World Cup Tickets API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches/{id}/tickets:
    get:
      operationId: listMatchTickets
      tags: [tickets]
      summary: List tickets for a match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
  /tickets:
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create ticket (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Created ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update ticket (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    delete:
      operationId: deleteTicket
      tags: [tickets]
      summary: Delete ticket (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all posts/news
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Ticket:
      type: object
      properties:
        id:
          type: integer
        matchId:
          type: integer
        category:
          type: string
        categoryAr:
          type: string
        seatNumber:
          type: string
        section:
          type: string
        row:
          type: ["string", "null"]
        price:
          type: number
        status:
          type: string
          enum: [available, reserved, sold]
        createdAt:
          type: string
      required:
        - id
        - matchId
        - category
        - categoryAr
        - seatNumber
        - section
        - price
        - status
        - createdAt
    TicketInput:
      type: object
      properties:
        matchId:
          type: integer
        category:
          type: string
        categoryAr:
          type: string
        seatNumber:
          type: string
        section:
          type: string
        row:
          type: string
        price:
          type: number
      required:
        - matchId
        - category
        - categoryAr
        - seatNumber
        - section
        - price
    TicketUpdate:
      type: object
      properties:
        category:
          type: string
        categoryAr:
          type: string
        seatNumber:
          type: string
        section:
          type: string
        row:
          type: string
        price:
          type: number
        status:
          type: string
          enum: [available, reserved, sold]
...
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        content:
          type: string
        contentAr:
          type: string
        image:
          type: ["string", "null"]
        isPublished:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - titleAr
        - content
        - contentAr
        - isPublished
        - createdAt
    PostInput:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        content:
          type: string
        contentAr:
          type: string
        image:
          type: string
        isPublished:
          type: boolean
      required:
        - title
        - titleAr
        - content
        - contentAr
    PostUpdate:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        content:
          type: string
        contentAr:
          type: string
        image:
          type: string
        isPublished:
          type: boolean
...